fix: make it compatible with Symfony 4, 5 and 6
Since creating branches in the library is not an option, a custom
`DateTimeNormalizer` for Symfony 4 and 5 is injected accordingly.
This needs to be revisited when support for Symfony 4 will be
dropped. This will likely happen in Drupal 10.

Symfony serializer needs to be set to ^4.4.12 since it removes
the uses of `libxml_disable_entity_loader` which is triggering
an issue in PHP 8.